text2image_service/requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
celery==5.2.7
redis==4.5.1
requests==2.28.2
argon2-cffi==21.3.0
gunicorn==20.1.0
